# src/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CustomStatus:
    properties:
      original: {}
      statusInt:
        type: integer
      statusString:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
  models.RequestTask:
    properties:
      description:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  models.Task:
    properties:
      created_at:
        type: string
      decription:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/models.CustomStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления задачами (TODO-лист)
  title: TaskAPI
  version: "1.0"
paths:
  /api/v1/tasks/:
    get:
      consumes:
      - application/json
      description: Возвращает все задачи
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Добавляет задачу
      parameters:
      - description: JSON задачи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      tags:
      - tasks
  /api/v1/tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет задачу по id
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Изменяет задачу по id
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: JSON задачи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      tags:
      - tasks
swagger: "2.0"